+++ packages/app/components/EditMode/EditMode.tsx
@@ -3,8 +3,9 @@ import { Button } from 'semantic-ui-react';
 import { compose } from 'recompose';
 import styled from 'styled-components';
 
-import withAuth, { WithAuthProps } from '../../Login/decorators/withAuth';
-import withEditMode, { WithEditModeProps } from '../../decorators/withEditMode';
+import UploadButton from './UploadButton';
+import withAuth, { WithAuthProps } from '../Login/decorators/withAuth';
+import withEditMode, { WithEditModeProps } from '../decorators/withEditMode';
 
 const Wrapper = styled.div`
   left: 400px;
@@ -23,7 +24,8 @@ const Edit: React.SFC<WithAuthProps & WithEditModeProps> = ({
       {isEditMode ? (
         <span>
           <Button onClick={closeEditMode}>Cancel</Button>
-          <Button onClick={closeEditMode} primary>
+          <UploadButton />
+          <Button onClick={closeEditMode} primary={true}>
             Save
           </Button>
         </span>
@@ -32,7 +34,7 @@ const Edit: React.SFC<WithAuthProps & WithEditModeProps> = ({
           content="Edit this territory"
           icon="edit"
           onClick={openEditMode}
-          primary
+          primary={true}
         />
       )}
     </Wrapper>
