+++ packages/app/components/EditMode/UploadButton/index.ts
@@ -1,0 +1,3 @@
+import UploadButton from './UploadButton';
+
+export default UploadButton;
