+++ packages/app/components/EditMode/index.ts
@@ -1,0 +1,3 @@
+import EditMode from './EditMode';
+
+export default EditMode;
