--- packages/app/components/MainMap/Edit/index.ts
@@ -1,3 +1,0 @@
-import Edit from './Edit';
-
-export default Edit;
