+++ packages/app/pages/map.tsx
@@ -14,7 +14,7 @@ import { withEditModeStore } from '../components/decorators/withEditMode';
 // @see https://github.com/zeit/next.js/issues/4515
 const MainMap = dynamic(import('../components/MainMap'), { ssr: false });
 // @ts-ignore
-const Edit = dynamic(import('../components/MainMap/Edit'), { ssr: false });
+const EditMode = dynamic(import('../components/EditMode'), { ssr: false });
 // @ts-ignore
 const Login = dynamic(import('../components/Login'), { ssr: false });
 
@@ -23,7 +23,7 @@ const Map = () => (
     <LeftPane>
       <MainMap />
     </LeftPane>
-    <Edit />
+    <EditMode />
     <Login />
   </Provider>
 );
